requirements.txt
# requirements.txt - Required packages

# Core packages
numpy>=1.20.0
opencv-python>=4.5.0
scipy>=1.7.0

# Deep learning
torch>=2.0.0
ultralytics>=8.0.0  # YOLOv8

# Web interface
flask>=2.0.0

# Utilities
PyYAML>=6.0
pillow>=9.0.0
pandas>=1.3.0
matplotlib>=3.4.0
